+++ generators/app/templates/config/services/cipher.js
@@ -1,0 +1,5 @@
+module.exports.services.cipher = {
+  jwt: {
+    secretKey: "<%= answers['application:secret'] %>"
+  }
+};